+++ src/features/background/interfaces/pixabay/IPixabayVideo.ts
@@ -1,6 +1,14 @@
-export interface IPixabayVideo {
+type PixabayVideoSize = {
 	url: string;
 	width: number;
 	height: number;
 	size: number;
+	thumbnail: string;
+};
+
+export interface IPixabayVideo {
+	// large: PixabayVideoSize;
+	medium: PixabayVideoSize;
+	small: PixabayVideoSize;
+	tiny: PixabayVideoSize;
 }