+++ src/features/background/state/backgroundMedia.ts
@@ -1,15 +1,11 @@
 import { createContext, Dispatch, SetStateAction } from "react";
 
-import { IPexelsPhoto, IPixabayVideo } from "@/features/background/interfaces";
+import { IPixabayVideo } from "@/features/background/interfaces";
 
 interface BackgroundMediaContextProps {
-	images: IPexelsPhoto[];
 	videos: IPixabayVideo[];
-	backgroundQuery: string;
 	setBackgroundMedia: Dispatch<SetStateAction<string>>;
-	setImages: Dispatch<SetStateAction<IPexelsPhoto[]>>;
 	setVideos: Dispatch<SetStateAction<IPixabayVideo[]>>;
-	setBackgroundQuery: Dispatch<SetStateAction<string>>;
 	setCountdownFontColor: Dispatch<SetStateAction<string>>;
 }